+++ src/main/scala/nasa/nccs/edas/sources/cdscan.scala
@@ -49,8 +49,8 @@ class DatasetFileHeaders(val aggDim: String, val aggFileMap: Seq[FileHeader]) {
     assert( aggFileMap.nonEmpty, "Error, aggregated dataset has no files!")
     aggFileMap.head.nElem
   }
-  def getAggAxisValues: Array[Float] =
-    aggFileMap.foldLeft(Array[Float]()) { _ ++ _.axisValues }
+  def getAggAxisValues: Array[Int] =
+    aggFileMap.foldLeft(Array[Int]()) { _ ++ _.axisValues }
 }
 
 
@@ -130,10 +130,10 @@ object FileHeader extends Loggable {
     }
   }
 
-  def getTimeValues(ncDataset: NetcdfDataset, coordAxis: VariableDS, start_index: Int = 0, end_index: Int = -1, stride: Int = 1): ( Array[Float], Array[Array[Float]] ) = {
+  def getTimeValues(ncDataset: NetcdfDataset, coordAxis: VariableDS, start_index: Int = 0, end_index: Int = -1, stride: Int = 1): ( Array[Int], Array[Array[Int]] ) = {
     val timeAxis: CoordinateAxis1DTime = CoordinateAxis1DTime.factory(ncDataset, coordAxis, new Formatter())
     val timeCalValues: List[CalendarDate] = timeAxis.getCalendarDates.toList
-    val bounds: Array[Array[Float]] = ((0 until timeAxis.getShape(0)) map (index => timeAxis.getCoordBoundsDate(index) map ( cdsutils.toValue ))).toArray
+    val bounds: Array[Array[Int]] = ((0 until timeAxis.getShape(0)) map (index => timeAxis.getCoordBoundsDate(index) map ( cdsutils.toValue ))).toArray
     val timeValues = timeCalValues.map( cdsutils.toValue ).toArray
     val datesSample = timeCalValues.subList(0,5)
     val timeValuesSample = timeValues.slice(0,5)
@@ -142,7 +142,7 @@ object FileHeader extends Loggable {
   }
 
 
-  def getTimeCoordValues(ncDataset: NetcdfDataset): ( Array[Float], Array[Array[Float]] ) = {
+  def getTimeCoordValues(ncDataset: NetcdfDataset): ( Array[Int], Array[Array[Int]] ) = {
     val result = Option(ncDataset.findCoordinateAxis(AxisType.Time)) match {
       case Some(timeAxis) => getTimeValues(ncDataset, timeAxis)
       case None => throw new Exception( "ncDataset does not have a time axis: " + ncDataset.getReferencedFile.getLocation )
@@ -152,14 +152,14 @@ object FileHeader extends Loggable {
 }
 
 class FileHeader(val filePath: String,
-                 val axisValues: Array[Float],
-                 val boundsValues: Array[Array[Float]],
+                 val axisValues: Array[Int],
+                 val boundsValues: Array[Array[Int]],
                  val timeRegular: Boolean,
                  val varNames: List[String],
                  val coordVarNames: List[String]
                 ) {
   def nElem: Int = axisValues.length
-  def startValue: Float = boundsValues.head(0)
+  def startValue: Int = boundsValues.head(0)
   def endValue: Float = boundsValues.last(1)
   def dt = ( endValue + 1 - startValue ) / boundsValues.length
   def startDate: String = cdsutils.toDate(startValue).toString
